# packages/openapi/api.yaml
components:
  schemas:
...
    FileInfo:
      properties:
        children:
          default: []
          items:
            $ref: '#/components/schemas/FileInfo'
          type: array
        id:
          type: string
        isDirectory:
          type: boolean
        isMarimoFile:
          type: boolean
        lastModified:
          anyOf:
          - type: number
          - type: 'null'
          default: null
        name:
          type: string
        opengraph:
          anyOf:
          - type: 'null'
          - $ref: '#/components/schemas/OpenGraphMetadata'
          default: null
        path:
          type: string
      required:
      - id
      - path
      - name
      - isDirectory
      - isMarimoFile
      title: FileInfo
      type: object
...
    MarimoFile:
      properties:
        initializationId:
          anyOf:
          - type: string
          - type: 'null'
          default: null
        lastModified:
          anyOf:
          - type: number
          - type: 'null'
          default: null
        name:
          type: string
        path:
          type: string
        sessionId:
          anyOf:
          - $ref: '#/components/schemas/SessionId'
          - type: 'null'
          default: null
      required:
      - name
      - path
      title: MarimoFile
      type: object
...
    OpenGraphMetadata:
      description: "OpenGraph-style metadata for a notebook.\n\n    The `image` field\
        \ may be either:\n    - a relative path (typically under `__marimo__/`), or\n\
        \    - an absolute HTTPS URL."
      properties:
        description:
          anyOf:
          - type: string
          - type: 'null'
          default: null
        image:
          anyOf:
          - type: string
          - type: 'null'
          default: null
        title:
          anyOf:
          - type: string
          - type: 'null'
          default: null
      required: []
      title: OpenGraphMetadata
      type: object
    OpenTemplateRequest:
      properties:
        templatePath:
          type: string
      required:
      - templatePath
      title: OpenTemplateRequest
      type: object
    OpenTutorialRequest:
      properties:
        tutorialId:
          anyOf:
          - enum:
            - dataflow
            - external-dependencies
            - fileformat
            - for-jupyter-users
            - intro
            - layout
            - markdown
            - plots
            - sql
            - ui
          - enum:
            - markdown-format
      required:
      - tutorialId
      title: OpenTutorialRequest
      type: object
...
    RunningNotebooksResponse:
      properties:
        files:
          items:
            $ref: '#/components/schemas/MarimoFile'
          type: array
      required:
      - files
      title: RunningNotebooksResponse
      type: object
...
    SessionId:
      format: session-id
      type: string
...
    ShutdownSessionRequest:
      properties:
        sessionId:
          $ref: '#/components/schemas/SessionId'
      required:
      - sessionId
      title: ShutdownSessionRequest
      type: object
...
    WorkspaceFilesRequest:
      properties:
        includeMarkdown:
          default: false
          type: boolean
      required: []
      title: WorkspaceFilesRequest
      type: object
    WorkspaceFilesResponse:
      properties:
        fileCount:
          default: 0
          type: integer
        files:
          items:
            $ref: '#/components/schemas/FileInfo'
          type: array
        hasMore:
          default: false
          type: boolean
        root:
          type: string
      required:
      - root
      - files
      title: WorkspaceFilesResponse
      type: object
...
info:
  title: marimo API
openapi: 3.1.0
paths:
...
  /api/home/shutdown_session:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShutdownSessionRequest'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunningNotebooksResponse'
          description: Shutdown the current session
  /api/home/template/open:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenTemplateRequest'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarimoFile'
          description: Open a template notebook
...
  /api/home/tutorial/open:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenTutorialRequest'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarimoFile'
          description: Open a new tutorial
  /api/home/workspace_files:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkspaceFilesRequest'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceFilesResponse'
          description: Get the files in the workspace
...